Refactored dialog classes, added restyled progress bar dialog
